+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.4] — 2026-09-12
 
+* fix(bridge): export missing `initRustLib` and `disposeRustLib` for `ProjectManager` classes.
 * chore: updated dependencies.
 
 ## [1.0.3] — 2026-09-06